Test: Colocate Browser Mode pilots
