+++ packages/components/package.json
@@ -115,7 +115,9 @@
 		"@wordpress/stylelint-tools": "file:../../tools/stylelint",
 		"snapshot-diff": "^0.10.0",
 		"storybook": "^10.4.3",
-		"timezone-mock": "^1.3.6"
+		"timezone-mock": "^1.3.6",
+		"vitest": "^4.1.10",
+		"vitest-browser-react": "^2.2.0"
 	},
 	"peerDependencies": {
 		"@types/react": "^18 || ^19",
