+++ packages/components/src/composite/test/browser-mode-pilot.browser.test.tsx
@@ -8,8 +8,7 @@ import { render } from 'vitest-browser-react';
 /**
  * Internal dependencies
  */
-import { Composite } from '../../../packages/components/src/composite';
-import Tooltip from '../../../packages/components/src/tooltip';
+import { Composite } from '..';
 
 test( 'moves focus through a composite with real keyboard input', async () => {
 	const user = userEvent.setup();
@@ -40,28 +39,4 @@ test( 'moves focus through a composite with real keyboard input', async () => {
 
 	await user.tab();
 	await expect.element( after ).toHaveFocus();
-} );
-
-test( 'shows a portal tooltip through real pointer input', async () => {
-	const user = userEvent.setup();
-	const view = await render(
-		<Tooltip delay={ 0 } text="Browser mode tooltip">
-			<button>Tooltip anchor</button>
-		</Tooltip>
-	);
-	const anchor = page.getByRole( 'button', {
-		name: 'Tooltip anchor',
-	} );
-	const tooltip = page.getByRole( 'tooltip', {
-		name: 'Browser mode tooltip',
-	} );
-
-	await expect.element( tooltip ).not.toBeInTheDocument();
-	await user.hover( anchor );
-	await expect.element( tooltip ).toBeVisible();
-	await expect
-		.element( anchor )
-		.toHaveAttribute( 'aria-describedby', tooltip.element().id );
-
-	expect( view.container.contains( tooltip.element() ) ).toBe( false );
 } );
